# apps/web/openapi/mobile-v1.json
{
  "openapi": "3.1.0",
  "info": { "title": "Kestrel One Mobile API", "version": "1.1.0", "description": "Authenticated companion contract. Projects are read-only and all execution state is hosted." },
  "servers": [{ "url": "/api/mobile/v1" }],
  "security": [{ "sessionCookie": [] }, { "sessionBearer": [] }],
  "paths": {
...
    "/threads/{id}": { "get": { "operationId": "getMobileThreadSnapshot", "parameters": [{ "$ref": "#/components/parameters/Id" }], "responses": { "200": { "$ref": "#/components/responses/ThreadSnapshot" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } } },
...
  },
  "components": {
    "securitySchemes": { "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "better-auth.session_token" }, "sessionBearer": { "type": "http", "scheme": "bearer" } },
    "parameters": {
      "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 200 } },
...
    },
    "responses": {
      "Error": { "description": "Stable public failure.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
      "ThreadSnapshot": { "description": "Authoritative Thread snapshot.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ThreadSnapshot" } } } },
...
    },
    "schemas": {
      "IsoDate": { "type": "string", "format": "date-time" },
      "TurnStatus": { "type": "string", "enum": ["queued", "running", "waiting_for_input", "completed", "failed", "cancelled"] },
...
      "ErrorResponse": { "type": "object", "required": ["error"], "properties": { "error": { "type": "object", "required": ["code", "message", "retryable"], "properties": { "code": { "type": "string", "enum": ["BAD_REQUEST", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "CONFLICT", "SERVICE_UNAVAILABLE", "INTERNAL_ERROR", "ORGANIZATION_SETUP_REQUIRED"] }, "message": { "type": "string" }, "retryable": { "type": "boolean" }, "nextStep": { "type": "string", "enum": ["model_access", "workspace_compute", "environment_execution"] }, "settingsPath": { "type": "string", "enum": ["/settings/organization/ai-providers", "/settings/organization/setup"] } }, "additionalProperties": false } }, "additionalProperties": false },
...
      "MobileThread": { "type": "object", "required": ["id", "title", "project", "preview", "runStatus", "createdAt", "updatedAt"], "properties": { "id": { "type": "string" }, "title": { "type": "string" }, "project": { "oneOf": [{ "$ref": "#/components/schemas/ProjectRef" }, { "type": "null" }] }, "preview": { "type": "string" }, "runStatus": { "oneOf": [{ "$ref": "#/components/schemas/TurnStatus" }, { "type": "null" }] }, "createdAt": { "$ref": "#/components/schemas/IsoDate" }, "updatedAt": { "$ref": "#/components/schemas/IsoDate" } }, "additionalProperties": false },
      "ProjectRef": { "type": "object", "required": ["id", "name"], "properties": { "id": { "type": "string" }, "name": { "type": "string" } }, "additionalProperties": false },
      "TextPart": { "type": "object", "required": ["type", "text"], "properties": { "type": { "type": "string", "const": "text" }, "text": { "type": "string" } }, "additionalProperties": false },
      "SourceUrlPart": { "type": "object", "required": ["type", "sourceId", "url", "title"], "properties": { "type": { "type": "string", "const": "source_url" }, "sourceId": { "type": "string" }, "url": { "type": "string", "format": "uri" }, "title": { "type": ["string", "null"] } }, "additionalProperties": false },
      "SourceDocumentPart": { "type": "object", "required": ["type", "sourceId", "mediaType", "title", "filename"], "properties": { "type": { "type": "string", "const": "source_document" }, "sourceId": { "type": "string" }, "mediaType": { "type": "string" }, "title": { "type": "string" }, "filename": { "type": ["string", "null"] } }, "additionalProperties": false },
      "ToolStatusPart": { "type": "object", "required": ["type", "toolCallId", "toolName", "state"], "properties": { "type": { "type": "string", "const": "tool_status" }, "toolCallId": { "type": "string" }, "toolName": { "type": "string" }, "state": { "type": "string", "enum": ["pending", "running", "waiting_for_approval", "completed", "failed", "denied", "unavailable"] } }, "additionalProperties": false },
      "ProgressPart": { "type": "object", "required": ["type", "id", "category", "label", "text", "timestamp"], "properties": { "type": { "type": "string", "const": "progress" }, "id": { "type": "string" }, "category": { "type": "string", "enum": ["runtime", "agent"] }, "label": { "type": "string" }, "text": { "type": "string" }, "timestamp": { "$ref": "#/components/schemas/IsoDate" } }, "additionalProperties": false },
      "CitationPart": { "type": "object", "required": ["type", "id", "title", "url", "documentId", "excerpt"], "properties": { "type": { "type": "string", "const": "citation" }, "id": { "type": "string" }, "title": { "type": "string" }, "url": { "type": ["string", "null"] }, "documentId": { "type": ["string", "null"] }, "excerpt": { "type": ["string", "null"] } }, "additionalProperties": false },
      "ArtifactPart": { "type": "object", "required": ["type", "id", "title", "kind", "url", "mediaType", "previewId", "sizeBytes", "fileCount", "expiresAt", "warning"], "properties": { "type": { "type": "string", "const": "artifact" }, "id": { "type": "string" }, "title": { "type": "string" }, "kind": { "type": "string" }, "url": { "type": ["string", "null"] }, "mediaType": { "type": ["string", "null"] }, "previewId": { "type": ["string", "null"] }, "sizeBytes": { "type": ["number", "null"], "minimum": 0 }, "fileCount": { "type": ["integer", "null"], "minimum": 1 }, "expiresAt": { "oneOf": [{ "$ref": "#/components/schemas/IsoDate" }, { "type": "null" }] }, "warning": { "type": ["string", "null"] } }, "additionalProperties": false },
      "InteractionStatusPart": { "type": "object", "required": ["type", "requestId", "kind", "prompt", "status"], "properties": { "type": { "type": "string", "const": "interaction_status" }, "requestId": { "type": "string" }, "kind": { "type": "string", "enum": ["question", "approval"] }, "prompt": { "type": "string" }, "status": { "type": "string", "enum": ["pending", "resolved", "cancelled"] } }, "additionalProperties": false },
      "AssistantStatusPart": { "type": "object", "required": ["type", "status", "errorCode", "message"], "properties": { "type": { "type": "string", "const": "assistant_status" }, "status": { "type": "string", "enum": ["working", "completed", "waiting", "failed", "cancelled", "contract_failure"] }, "errorCode": { "oneOf": [{ "type": "string", "enum": ["AGENT_RUN_FAILED", "AGENT_RUN_CANCELLED", "PRESENTATION_CONTRACT_FAILURE"] }, { "type": "null" }] }, "message": { "type": ["string", "null"] } }, "additionalProperties": false },
      "MessagePart": { "oneOf": [{ "$ref": "#/components/schemas/TextPart" }, { "$ref": "#/components/schemas/SourceUrlPart" }, { "$ref": "#/components/schemas/SourceDocumentPart" }, { "$ref": "#/components/schemas/ToolStatusPart" }, { "$ref": "#/components/schemas/ProgressPart" }, { "$ref": "#/components/schemas/CitationPart" }, { "$ref": "#/components/schemas/ArtifactPart" }, { "$ref": "#/components/schemas/InteractionStatusPart" }, { "$ref": "#/components/schemas/AssistantStatusPart" }] },
      "MobileMessage": { "type": "object", "required": ["id", "turnId", "role", "parts", "createdAt"], "properties": { "id": { "type": "string" }, "turnId": { "type": ["string", "null"] }, "role": { "type": "string", "enum": ["user", "assistant", "system"] }, "parts": { "type": "array", "items": { "$ref": "#/components/schemas/MessagePart" } }, "createdAt": { "$ref": "#/components/schemas/IsoDate" } }, "additionalProperties": false },
      "TurnFailure": { "type": "object", "required": ["code", "message", "retryable"], "properties": { "code": { "type": "string", "enum": ["AGENT_RUN_FAILED", "AGENT_RUN_CANCELLED"] }, "message": { "type": "string" }, "retryable": { "type": "boolean" } }, "additionalProperties": false },
      "MobileTurn": { "type": "object", "required": ["id", "threadId", "inputMessageId", "sequence", "status", "failure", "cancelRequestedAt", "startedAt", "finishedAt", "createdAt", "updatedAt"], "properties": { "id": { "type": "string" }, "threadId": { "type": "string" }, "inputMessageId": { "type": ["string", "null"] }, "sequence": { "type": "integer" }, "status": { "$ref": "#/components/schemas/TurnStatus" }, "failure": { "oneOf": [{ "$ref": "#/components/schemas/TurnFailure" }, { "type": "null" }] }, "cancelRequestedAt": { "oneOf": [{ "$ref": "#/components/schemas/IsoDate" }, { "type": "null" }] }, "startedAt": { "oneOf": [{ "$ref": "#/components/schemas/IsoDate" }, { "type": "null" }] }, "finishedAt": { "oneOf": [{ "$ref": "#/components/schemas/IsoDate" }, { "type": "null" }] }, "createdAt": { "$ref": "#/components/schemas/IsoDate" }, "updatedAt": { "$ref": "#/components/schemas/IsoDate" } }, "additionalProperties": false },
      "ThreadQueue": { "type": "object", "required": ["state", "pauseReason", "activeTurnId", "version", "orderedTurnIds"], "properties": { "state": { "type": "string", "enum": ["running", "paused"] }, "pauseReason": { "type": ["string", "null"] }, "activeTurnId": { "type": ["string", "null"] }, "version": { "type": "integer", "minimum": 0 }, "orderedTurnIds": { "type": "array", "items": { "type": "string" } } }, "additionalProperties": false },
      "InteractionField": { "type": "object", "required": ["name", "label", "type", "required"], "properties": { "name": { "type": "string" }, "label": { "type": "string" }, "type": { "type": "string", "enum": ["text", "number", "boolean", "select", "multi_select"] }, "required": { "type": "boolean" }, "options": { "type": "array", "items": { "type": "string" } } }, "additionalProperties": false },
      "MobileInteraction": { "type": "object", "required": ["id", "kind", "title", "prompt", "fields", "createdAt"], "properties": { "id": { "type": "string" }, "kind": { "type": "string", "enum": ["question", "approval"] }, "version": { "type": "string", "enum": ["legacy", "runner_hosted_tool_approval_interaction_v4"] }, "decisions": { "type": "array", "items": { "type": "string", "enum": ["approve", "deny", "decline", "approve_once", "remember_approval"] } }, "title": { "type": "string" }, "prompt": { "type": "string" }, "fields": { "type": "array", "items": { "$ref": "#/components/schemas/InteractionField" } }, "createdAt": { "$ref": "#/components/schemas/IsoDate" } }, "additionalProperties": false },
      "ThreadSnapshot": { "type": "object", "required": ["snapshotVersion", "thread", "messages", "turns", "queue", "interactions"], "properties": { "snapshotVersion": { "type": "string" }, "thread": { "$ref": "#/components/schemas/MobileThread" }, "messages": { "type": "array", "items": { "$ref": "#/components/schemas/MobileMessage" } }, "turns": { "type": "array", "items": { "$ref": "#/components/schemas/MobileTurn" } }, "queue": { "$ref": "#/components/schemas/ThreadQueue" }, "interactions": { "type": "array", "items": { "$ref": "#/components/schemas/MobileInteraction" } } }, "additionalProperties": false },
...
    }
  },
...
}